+++ reana_client/cli/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # This file is part of REANA.
-# Copyright (C) 2017, 2018, 2019, 2020, 2021 CERN.
+# Copyright (C) 2017, 2018, 2019, 2020, 2021, 2022 CERN.
 #
 # REANA is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
@@ -13,7 +13,7 @@ import sys
 import click
 from urllib3 import disable_warnings
 
-from reana_client.cli import workflow, files, ping, secrets, quotas
+from reana_client.cli import workflow, files, ping, secrets, quotas, retention_rules
 from reana_client.utils import get_api_url
 
 DEBUG_LOG_FORMAT = (
@@ -43,6 +43,7 @@ class ReanaCLI(click.Group):
         workflow.workflow_execution_group,
         workflow.interactive_group,
         files.files_group,
+        retention_rules.retention_rules_group,
         secrets.secrets_group,
     ]